requirements.txt
annoy==1.17.3
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
Django==6.0.3
django-bootstrap5==26.2
django-cors-headers==4.9.0
django-ratelimit==4.1.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
idna==3.11
joblib==1.5.3
numpy==2.4.3
pandas==3.0.1
pip==25.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.3.0
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
spotipy==2.26.0
sqlparse==0.5.5
threadpoolctl==3.6.0
tzdata==2025.3
urllib3==2.6.3